Move time converter to Media package and add tests.
