+++ packages/media/src/index.js
@@ -24,6 +24,7 @@ export { default as formatDuration } from './formatDuration';
 export { default as createFileReader } from './createFileReader';
 export { default as fetchRemoteFile } from './fetchRemoteFile';
 export { default as fetchRemoteBlob } from './fetchRemoteBlob';
+export { default as formatMsToHMS } from './formatMsToHMS';
 export { default as getFileNameFromUrl } from './getFileNameFromUrl';
 export { default as getFileNameWithExt } from './getFileNameWithExt';
 export { default as getExtensionFromMimeType } from './getExtensionFromMimeType';
